requirements.txt
streamlit>=1.40.0
opencv-python>=4.8.0
numpy>=1.24.0
face-recognition>=1.3.0
Pillow>=10.0.0
pandas>=2.0.0
python-dotenv>=1.0.0